+++ HISTORY.md
@@ -1,6 +1,7 @@
 unreleased
 ==========
 
+  * Support `If-Match` and `If-Unmodified-Since` headers
   * Add `res` and `path` arguments to `directory` event
   * Remove usage of `res._headers` private field
     - Improves compatibility with Node.js 8 nightly